backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.10.1
pydantic-settings==2.6.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.1
python-multipart==0.0.17
pytest==8.3.3
httpx==0.27.2
pip-audit==2.7.3
numpy==2.2.1
joblib==1.4.2
torch>=2.6.0
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10
